Converted mobile offense flags to BitVector.
struct race_type's offense flags will now be stored as a string.
Area files now have a proper version number to allow for more controlled changes to the area file format.
Fixed a bug in get_skill() where checks for ACT_WARRIOR and ACT_THIEF were made against the off_flags bitvector rather than the correct act bitvector.

+++ src/save.c
@@ -34,7 +34,7 @@ extern bool chaos;
 
 
 /* Convert a standard bitvector to our new BitVector type */
-void bv_from_old_bitvector(BitVector *bv, long old_bitvector)
+void bv_from_old_bitvector(BitVector *bv, uint32_t old_bitvector)
 {
     int index = 0;
